fix: use explicit default from prompts (#303)
closes #171

I think the issue is that `esModuleInterop` TS compiler setting forces consumers to be as well. And that is not good, since its an ugly detail to interop with legacy non-es-modules. We do not want `santa` to be cause for user to have to turn that compiler option on in their project.