fix: debug module types
closes #208

We need to keep debug module types dep in production deps becuase we
export its types for consumers. When it was a dev dep it would not be
installed for consumers leading to type errors for them.

For a similar case of issue refer to: https://github.com/prisma-labs/nexus-prisma/issues/571