fix: forward prompts type w/o also esModuleInterop
Revisiting 3044faad715b6743843c14155173532f39e3bc16

The previous commit wasn't right. For one thing esModuleInterop was
enabled on the base TS config file so it was never actually removed. But
moreover the problem is not its use per-se, the problem is that our app
was forwarding a type that was imported requiring esModuleInterop. By
having the consumers exposed to a type that required esModuleInterop for
us so to it will for them.

This commit takes a different solution. It re-creates the type we wish
to export. By doing this, the fact we import prompts via esModuleInterop
is decoupled from how the consumer wants/must configure their TS config.