refactror: settings lib
This new library begins a principaled settings system.
We will be able to add systematic features on top of this.
Like environment variable consumption.
And value source tracking.

The primary motiviation right now is value source tracking. We need it
in order to implement subscriptions nicely. We need to know if the user
has subscriptions in the schema but disabled server subscriptions, or
vice-versa.